+++ src/rex.formbuilder/NEWS.rst
@@ -10,6 +10,7 @@ REX.FORMBUILDER Change Log
 
 * Added  ``formbuilder-pick-draft`` and ``formbuilder-edit-draft`` action types
   for use in rex.action wizards.
+* Removed inappropriate properties from Boolean field.
 
 
 5.7.0 (2016-04-22)